Move all tests into a single file
