+++ serde_arrow/src/internal/schema/serde/mod.rs
@@ -2,6 +2,3 @@
 //!
 pub mod deserialize;
 pub mod serialize;
-
-#[cfg(test)]
-mod test;
